requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-braces==1.13.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
whitenoise==4.1.4
